+++ lib/__tests__/start.test.js
@@ -27,14 +27,14 @@ jest.mock('process');
 
 console.log = jest.fn();
 
-const serverController = require('../serverController.js');
+const start = require('../start.js');
 
 beforeEach(() => jest.resetAllMocks());
 
 describe('start live reload', () => {
   test('uses inital port 35729', () => {
     findUnusedPort.mockResolvedValue();
-    serverController.startLiveReloadServer();
+    start.startLiveReloadServer();
     expect(findUnusedPort).toHaveBeenCalledWith(35729);
   });
 
@@ -42,7 +42,7 @@ describe('start live reload', () => {
     expect.assertions(1);
     const unusedPort = 1234;
     findUnusedPort.mockResolvedValue(unusedPort);
-    return serverController.startLiveReloadServer().then(() => {
+    return start.startLiveReloadServer().then(() => {
       expect(liveReloadServer.start).toHaveBeenCalledWith(unusedPort);
     });
   });
@@ -51,7 +51,7 @@ describe('start live reload', () => {
     expect.assertions(1);
     const unusedPortError = new Error('no unused port');
     findUnusedPort.mockRejectedValue(unusedPortError);
-    return expect(serverController.startLiveReloadServer()).rejects.toEqual(
+    return expect(start.startLiveReloadServer()).rejects.toEqual(
       unusedPortError
     );
   });
@@ -62,7 +62,7 @@ describe('start server', () => {
     const customPort = 1234;
     program.port = customPort;
     findUnusedPort.mockResolvedValue();
-    serverController.startServer();
+    start.startServer();
     expect(findUnusedPort).toBeCalledWith(customPort);
     delete program.port;
   });
@@ -71,14 +71,14 @@ describe('start server', () => {
     const customPort = '4321';
     process.env.PORT = customPort;
     findUnusedPort.mockResolvedValue();
-    serverController.startServer();
+    start.startServer();
     expect(findUnusedPort).toBeCalledWith(customPort);
     delete process.env.PORT;
   });
 
   test('when no custom port specified, uses port 3000', () => {
     findUnusedPort.mockResolvedValue();
-    serverController.startServer();
+    start.startServer();
     expect(findUnusedPort).toBeCalledWith(3000);
   });
 
@@ -86,7 +86,7 @@ describe('start server', () => {
     expect.assertions(1);
     const port = 1357;
     findUnusedPort.mockResolvedValue(port);
-    return serverController.startServer().then(() => {
+    return start.startServer().then(() => {
       expect(server).toHaveBeenCalledWith(port);
     });
   });
@@ -98,7 +98,7 @@ describe('start server', () => {
     const port = 2468;
     findUnusedPort.mockResolvedValue(port);
     const expectedServerAddress = `http://localhost:${port}${baseUrl}`;
-    return serverController.startServer().then(() => {
+    return start.startServer().then(() => {
       expect(openBrowser).toHaveBeenCalledWith(expectedServerAddress);
     });
   });
@@ -109,7 +109,7 @@ describe('start docusaurus', () => {
     expect.assertions(1);
     program.watch = true;
     findUnusedPort.mockResolvedValue();
-    return serverController.startDocusaurus().then(() => {
+    return start.startDocusaurus().then(() => {
       expect(liveReloadServer.start).toBeCalled();
     });
   });
@@ -121,7 +121,7 @@ describe('start docusaurus', () => {
     findUnusedPort
       .mockRejectedValueOnce('could not find live reload port')
       .mockResolvedValueOnce();
-    return serverController.startDocusaurus().then(() => {
+    return start.startDocusaurus().then(() => {
       expect(server).toBeCalled();
     });
   });
@@ -130,7 +130,7 @@ describe('start docusaurus', () => {
     expect.assertions(2);
     program.watch = false;
     findUnusedPort.mockResolvedValue();
-    return serverController.startDocusaurus().then(() => {
+    return start.startDocusaurus().then(() => {
       expect(liveReloadServer.start).not.toBeCalled();
       expect(server).toBeCalled();
     });
