+++ lib/start-server.js
@@ -25,7 +25,7 @@ const program = require('commander');
 const CWD = process.cwd();
 const env = require('./server/env.js');
 
-const serverController = require('./serverController.js');
+const start = require('./start.js');
 
 if (!fs.existsSync(`${CWD}/siteConfig.js`)) {
   console.error(
@@ -44,7 +44,7 @@ program
   .option('--no-watch', 'Toggle live reload file watching')
   .parse(process.argv);
 
-serverController.startDocusaurus().catch(ex => {
+start.startDocusaurus().catch(ex => {
   console.error(`failed to start Docusaurus server: ${ex}`);
   process.exit(1);
 });
